feat(testnets): Add Hoodi and Ephemery support for Nimbus-Nethermind CL/EL

+++ deploy-nimbus-nethermind.py
@@ -26,6 +26,8 @@ import zipfile
 import random
 import sys
 import platform
+import tempfile
+import yaml
 from consolemenu import *
 from consolemenu.items import *
 import argparse
@@ -43,7 +45,7 @@ def clear_screen():
 clear_screen()  # Call the function to clear the screen
 
 # Valid configurations
-valid_networks = ['MAINNET', 'HOLESKY', 'SEPOLIA']
+valid_networks = ['MAINNET','HOODI','EPHEMERY', 'HOLESKY', 'SEPOLIA']
 valid_exec_clients = ['NETHERMIND']
 valid_consensus_clients = ['NIMBUS']
 valid_install_configs = ['Solo Staking Node', 'Full Node Only', 'Lido CSM Staking Node', 'Lido CSM Validator Client Only' ,'Validator Client Only', 'Failover Staking Node']
@@ -65,10 +67,12 @@ FEE_RECIPIENT_ADDRESS=os.getenv('FEE_RECIPIENT_ADDRESS')
 MEV_MIN_BID=os.getenv('MEV_MIN_BID')
 CSM_FEE_RECIPIENT_ADDRESS_MAINNET=os.getenv('CSM_FEE_RECIPIENT_ADDRESS_MAINNET')
 CSM_FEE_RECIPIENT_ADDRESS_HOLESKY=os.getenv('CSM_FEE_RECIPIENT_ADDRESS_HOLESKY')
+CSM_FEE_RECIPIENT_ADDRESS_HOODI=os.getenv('CSM_FEE_RECIPIENT_ADDRESS_HOODI')
 CSM_GRAFFITI=os.getenv('CSM_GRAFFITI')
 CSM_MEV_MIN_BID=os.getenv('CSM_MEV_MIN_BID')
 CSM_WITHDRAWAL_ADDRESS_MAINNET=os.getenv('CSM_WITHDRAWAL_ADDRESS_MAINNET')
 CSM_WITHDRAWAL_ADDRESS_HOLESKY=os.getenv('CSM_WITHDRAWAL_ADDRESS_HOLESKY')
+CSM_WITHDRAWAL_ADDRESS_HOODI=os.getenv('CSM_WITHDRAWAL_ADDRESS_HOODI')
 
 # Create argparse options
 parser = argparse.ArgumentParser(description='Validator Install Options :: CoinCashew.com',formatter_class=argparse.ArgumentDefaultsHelpFormatter)
@@ -118,7 +122,7 @@ if not args.network and not args.skip_prompts:
     index = SelectionMenu.get_selection(valid_networks,title='Validator Install Quickstart :: CoinCashew.com',subtitle='Installs Nethermind EL / Nimbus BN / Nimbus VC / MEVboost\nSelect Ethereum network:')
 
     # Exit selected
-    if index == 3:
+    if index == 5:
         exit(0)
 
     # Set network
@@ -168,6 +172,12 @@ else:
           elif eth_network == "holesky":
               FEE_RECIPIENT_ADDRESS=CSM_FEE_RECIPIENT_ADDRESS_HOLESKY
               CSM_WITHDRAWAL_ADDRESS=CSM_WITHDRAWAL_ADDRESS_HOLESKY
+          elif eth_network == "hoodi":
+              FEE_RECIPIENT_ADDRESS=CSM_FEE_RECIPIENT_ADDRESS_HOODI
+              CSM_WITHDRAWAL_ADDRESS=CSM_WITHDRAWAL_ADDRESS_HOODI
+          elif eth_network == "ephemery":
+              FEE_RECIPIENT_ADDRESS=CSM_FEE_RECIPIENT_ADDRESS_HOLESKY
+              CSM_WITHDRAWAL_ADDRESS=CSM_WITHDRAWAL_ADDRESS_HOLESKY
           else:
             print(f'Unsupported Lido CSM Staking Node network: {eth_network}')
             exit(1)
@@ -182,6 +192,12 @@ else:
               FEE_RECIPIENT_ADDRESS=CSM_FEE_RECIPIENT_ADDRESS_MAINNET
               CSM_WITHDRAWAL_ADDRESS=CSM_WITHDRAWAL_ADDRESS_MAINNET
           elif eth_network == "holesky":
+              FEE_RECIPIENT_ADDRESS=CSM_FEE_RECIPIENT_ADDRESS_HOLESKY
+              CSM_WITHDRAWAL_ADDRESS=CSM_WITHDRAWAL_ADDRESS_HOLESKY
+          elif eth_network == "hoodi":
+              FEE_RECIPIENT_ADDRESS=CSM_FEE_RECIPIENT_ADDRESS_HOODI
+              CSM_WITHDRAWAL_ADDRESS=CSM_WITHDRAWAL_ADDRESS_HOODI
+          elif eth_network == "ephemery":
               FEE_RECIPIENT_ADDRESS=CSM_FEE_RECIPIENT_ADDRESS_HOLESKY
               CSM_WITHDRAWAL_ADDRESS=CSM_WITHDRAWAL_ADDRESS_HOLESKY
           else:
@@ -199,6 +215,10 @@ else:
           MEVBOOST_ENABLED=True
           VALIDATOR_ENABLED=False
           VALIDATOR_ONLY=False
+
+# Ephemery and hoodi override, turn off mevboost
+if eth_network == "ephemery" or eth_network == "hoodi":
+    MEVBOOST_ENABLED=False
 
 execution_client=""
 consensus_client=""
@@ -253,8 +273,6 @@ if VALIDATOR_ONLY and args.vc_only_bn_address is None and not args.skip_prompts:
 else:
     BN_ADDRESS=args.vc_only_bn_address
 
-
-
 if not args.skip_prompts:
     # Format confirmation message
     if install_config == "Solo Staking Node" or install_config == "Lido CSM Staking Node" or install_config == "Failover Staking Node":
@@ -274,6 +292,68 @@ if not args.skip_prompts:
         print(f'\nInstall cancelled by user. \n\nWhen ready, re-run install command:\npython3 {file_name}')
         exit(0)
 
+def setup_ephemery_network(genesis_repository):
+    testnet_dir = "/opt/ethpillar/testnet"
+
+    def get_github_release(repo):
+        url = f"https://api.github.com/repos/{repo}/releases/latest"
+        response = requests.get(url)
+        if response.status_code == 200:
+            data = json.loads(response.text)
+            return data.get('tag_name')
+        else:
+            return None
+
+    def download_genesis_release(genesis_release):
+        # remove old genesis and setup dir
+        if os.path.exists(testnet_dir):
+            subprocess.run([f'sudo rm -rf {testnet_dir}'], shell=True)
+        subprocess.run([f'sudo mkdir -p {testnet_dir}'], shell=True)
+        subprocess.run([f'sudo chmod -R 755 {testnet_dir}'], shell=True)
+
+        # get latest genesis
+        url = f"https://github.com/{genesis_repository}/releases/download/{genesis_release}/testnet-all.tar.gz"
+        print(f">> Downloading {genesis_release} genesis files > URL: {url}")
+        response = requests.get(url, stream=True)
+        if response.status_code == 200:
+            temp_dir = tempfile.mkdtemp()
+            with tarfile.open(fileobj=response.raw, mode='r|gz') as tar:
+                tar.extractall(f"{temp_dir}")
+            os.system(f"sudo mv {temp_dir}/* {testnet_dir}")
+            print(f">> Successfully downloaded {genesis_release} genesis files")
+        else:
+            print("Failed to download genesis release")
+
+    genesis_release = get_github_release(genesis_repository)
+    if genesis_release:
+        download_genesis_release(genesis_release)
+    else:
+        print(f"Failed to retrieve genesis release for {genesis_repository}")
+
+def setup_hoodi_network():
+    testnet_dir = "/opt/ethpillar/testnet"
+    url = f"https://github.com/eth-clients/hoodi/archive/refs/tags/genesis.tar.gz"
+    ssz = f"https://github.com/eth-clients/hoodi/raw/refs/heads/main/metadata/genesis.ssz"
+
+    # remove old testnet_dir and setup dir
+    if os.path.exists(testnet_dir):
+        subprocess.run([f'sudo rm -rf {testnet_dir}'], shell=True)
+    subprocess.run([f'sudo mkdir -p {testnet_dir}'], shell=True)
+    subprocess.run([f'sudo chmod -R 755 {testnet_dir}'], shell=True)
+
+    # get genesis tar file
+    print(f">> Downloading HOODI genesis files > URL: {url}")
+    response = requests.get(url, stream=True)
+    if response.status_code == 200:
+        temp_dir = tempfile.mkdtemp()
+        with tarfile.open(fileobj=response.raw, mode='r|gz') as tar:
+            tar.extractall(f"{temp_dir}")
+        os.system(f"wget {ssz} -O {temp_dir}/hoodi-genesis/metadata/genesis.ssz")
+        os.system(f"sudo mv {temp_dir}/hoodi-genesis/metadata/* {testnet_dir}")
+        print(f">> Successfully downloaded HOODI genesis files")
+    else:
+        print("Failed to download genesis release")
+
 # Initialize sync urls for selected network
 if eth_network == "mainnet":
     sync_urls = mainnet_sync_urls
@@ -281,9 +361,18 @@ elif eth_network == "holesky":
     sync_urls = holesky_sync_urls
 elif eth_network == "sepolia":
     sync_urls = sepolia_sync_urls
+elif eth_network == "hoodi":
+    #TODO: remove post genesis
+    #sync_urls = hoodi_sync_urls
+    setup_hoodi_network()
+elif eth_network == "ephemery":
+    sync_urls = ephemery_sync_urls
+    setup_ephemery_network("ephemery-testnet/ephemery-genesis")
 
-# Use a random sync url
-sync_url = random.choice(sync_urls)[1]
+#TODO: remove post genesis
+if not eth_network == "hoodi":
+    # Use a random sync url
+    sync_url = random.choice(sync_urls)[1]
 
 def setup_node():
     if not VALIDATOR_ONLY:
@@ -385,6 +474,8 @@ def install_mevboost():
 
         if eth_network == 'mainnet':
             relay_options=mainnet_relay_options
+        elif eth_network == 'hoodi':
+            relay_options=hoodi_relay_options
         elif eth_network == 'holesky':
             relay_options=holesky_relay_options
         else:
@@ -490,6 +581,21 @@ def download_and_install_nethermind():
         # Remove the temporary zip file
         os.remove(temp_path)
 
+        # Process custom testnet configuration
+        if eth_network=="ephemery":
+            file_path = f"/opt/ethpillar/testnet/bootnode.txt"
+            with open(file_path, "r") as file:
+                bootnodes = ",".join(file.read().splitlines())
+            _network=f"--config none.json --Init.ChainSpecPath=/opt/ethpillar/testnet/chainspec.json --Discovery.Bootnodes={bootnodes} --JsonRpc.EnabledModules=Eth,Subscribe,Trace,TxPool,Web3,Personal,Proof,Net,Parity,Health,Rpc,Debug,Admin --JsonRpc.EngineHost=127.0.0.1 --JsonRpc.EnginePort=8551 --Init.IsMining=false"
+        elif eth_network=="hoodi":
+            file_path = f"/opt/ethpillar/testnet/enodes.yaml"
+            with open(file_path, "r") as file:
+                data = yaml.safe_load(file)
+            bootnodes = ','.join(data)
+            _network=f"--config none.json --Init.ChainSpecPath=/opt/ethpillar/testnet/chainspec.json --Discovery.Bootnodes={bootnodes} --JsonRpc.Enabled=true --JsonRpc.EnabledModules=Eth,Subscribe,Trace,TxPool,Web3,Personal,Proof,Net,Parity,Health,Rpc,Debug,Admin --JsonRpc.EngineHost=127.0.0.1 --JsonRpc.EnginePort=8551 --Init.IsMining=false"
+        else:
+            _network=f"--config {eth_network}"
+
         ##### NETHERMIND SERVICE FILE ###########
         nethermind_service_file = f'''[Unit]
 Description=Nethermind Execution Layer Client service for {eth_network.upper()}
@@ -507,7 +613,7 @@ KillSignal=SIGINT
 TimeoutStopSec=900
 WorkingDirectory=/var/lib/nethermind
 Environment="DOTNET_BUNDLE_EXTRACT_BASE_DIR=/var/lib/nethermind"
-ExecStart=/usr/local/bin/nethermind/nethermind --config {eth_network} --datadir="/var/lib/nethermind" --Network.DiscoveryPort {EL_P2P_PORT} --Network.P2PPort {EL_P2P_PORT} --Network.MaxActivePeers {EL_MAX_PEER_COUNT} --JsonRpc.Port {EL_RPC_PORT} --Metrics.Enabled true --Metrics.ExposePort 6060 --JsonRpc.JwtSecretFile {JWTSECRET_PATH} --Pruning.Mode=Hybrid --Pruning.FullPruningTrigger=VolumeFreeSpace --Pruning.FullPruningThresholdMb=300000
+ExecStart=/usr/local/bin/nethermind/nethermind {_network} --datadir="/var/lib/nethermind" --Network.DiscoveryPort {EL_P2P_PORT} --Network.P2PPort {EL_P2P_PORT} --Network.MaxActivePeers {EL_MAX_PEER_COUNT} --JsonRpc.Port {EL_RPC_PORT} --Metrics.Enabled true --Metrics.ExposePort 6060 --JsonRpc.JwtSecretFile {JWTSECRET_PATH} --Pruning.Mode=Hybrid --Pruning.FullPruningTrigger=VolumeFreeSpace --Pruning.FullPruningThresholdMb=300000
 
 [Install]
 WantedBy=multi-user.target
@@ -619,6 +725,21 @@ def install_nimbus():
         else:
             _feeparameters=''
 
+        # Process custom testnet configuration
+        if eth_network=="ephemery":
+            file_path = f"/opt/ethpillar/testnet/bootstrap_nodes.txt"
+            with open(file_path, "r") as file:
+                bootnodes = ",".join(file.read().splitlines())
+            _network=f"--network=/opt/ethpillar/testnet --bootstrap-node=${bootnodes}"
+        elif eth_network=="hoodi":
+            file_path = f"/opt/ethpillar/testnet/bootstrap_nodes.yaml"
+            with open(file_path, "r") as file:
+                data = yaml.safe_load(file)
+            bootnodes = ','.join(data)
+            _network=f"--network=/opt/ethpillar/testnet --bootstrap-node=${bootnodes}"
+        else:
+            _network=f"--network={eth_network}"
+
         ########### NIMBUS SERVICE FILE #############
         nimbus_service_file = f'''[Unit]
 Description=Nimbus Beacon Node Consensus Client service for {eth_network.upper()}
@@ -634,7 +755,7 @@ Restart=on-failure
 RestartSec=3
 KillSignal=SIGINT
 TimeoutStopSec=900
-ExecStart=/usr/local/bin/nimbus_beacon_node --network={eth_network} --data-dir=/var/lib/nimbus --tcp-port={CL_P2P_PORT} --udp-port={CL_P2P_PORT} --max-peers={CL_MAX_PEER_COUNT} --rest-port={CL_REST_PORT} --enr-auto-update=true --web3-url=http://127.0.0.1:8551 --rest --metrics --metrics-port=8008 --jwt-secret={JWTSECRET_PATH} --non-interactive --status-bar=false --in-process-validators=false {_feeparameters} {_mevparameters}
+ExecStart=/usr/local/bin/nimbus_beacon_node {_network} --data-dir=/var/lib/nimbus --tcp-port={CL_P2P_PORT} --udp-port={CL_P2P_PORT} --max-peers={CL_MAX_PEER_COUNT} --rest-port={CL_REST_PORT} --enr-auto-update=true --web3-url=http://127.0.0.1:8551 --rest --metrics --metrics-port=8008 --jwt-secret={JWTSECRET_PATH} --non-interactive --status-bar=false --in-process-validators=false {_feeparameters} {_mevparameters}
 
 [Install]
 WantedBy=multi-user.target
@@ -650,13 +771,23 @@ WantedBy=multi-user.target
         os.remove(nimbus_temp_file)
 
 def run_nimbus_checkpoint_sync():
+    # Skip for genesis event, TODO remove
+    if eth_network=="hoodi":
+        return
     if sync_url is not None and not VALIDATOR_ONLY:
         print(f'>> Running Checkpoint Sync. Using Sync URL: {sync_url}')
         db_path = "/var/lib/nimbus/db"
         os.system(f'sudo rm -rf {db_path}')
+
+        # Process custom testnet configuration
+        if eth_network=="hoodi" or eth_network=="ephemery":
+            _network=f"--network=/opt/ethpillar/testnet"
+        else:
+            _network=f"--network={eth_network}"
+
         subprocess.run([
             'sudo', '/usr/local/bin/nimbus_beacon_node', 'trustedNodeSync',
-            f'--network={eth_network}', '--data-dir=/var/lib/nimbus',
+            f'{_network}', '--data-dir=/var/lib/nimbus',
             f'--trusted-node-url={sync_url}', '--backfill=false'
         ])
         os.system(f'sudo chown -R consensus:consensus {db_path}')
@@ -733,8 +864,8 @@ def finish_install():
         print(f'Mevboost Version: \n{mevboost_version}\n')
 
     print(f'Network: {eth_network.upper()}\n')
-
-    if not VALIDATOR_ONLY:
+    #TODO remove for hoodi post genesis
+    if not VALIDATOR_ONLY and not eth_network=="hoodi":
         print(f'CheckPointSyncURL: {sync_url}\n')
 
     if VALIDATOR_ONLY and BN_ADDRESS:
